third_party/blink/web_tests/virtual/cross-partition-blob-url/external/wpt/storage-access-api/storage-access-beyond-cookies.thirdPartyBlobStorage.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies for third-party context accessing first-party Blob URLs
  assert_equals: expected "Third Party Blob URL tests completed successfully." but got "Cross-partition third-party blob URL was accessible after granting storage access"
Harness: the test ran to completion.
